Update melee.dm
About The Pull Request

Forged weapons are currently a little bit of a mess, with pierce type weapons being completely useless damage wise compared to other options, which have higher force, and wounding effects. This is likley due to piercing weapons historically having AP, but this got removed.

This request reduces the two hand multiplier on the two-hand axe to bring it a little closer to the other forged weapon options. While increasing the piercing weapon damage of both small and large, to be more similar to the other weapon catagories.

Pre-Merge Checklist

    [ X ] You tested this on a local server.
    [ X ] This code did not runtime during testing.
    [ X ] You documented all of your changes.

Changelog

🆑
tweak: Reduces the two hand axes multiplier to 1.5 (from 1.8) this still keeps it stronger than the Katana, but reduces it doing so much more damage than other forged weapons.
tweak: Increases base force of forged small pierce weapons (Dagger/Thorwing Knife) from 7 to 12 as they both lack AP or a wounding effect (Though knife can embed) this is less than the base force of weapons like the machete (15) which also has wound and can be stowed in a backpack. This allows them to be a little more competative without making them too powerful.
tweak: Increased the force of base force of forged largepierce weapons (Spear/Javelin) from 10 to 15. This put them on the same level as the large slashing weapons (Swords/Machete) without the additional wounding effect. Making them a possible option to use. This keeps spear the weakest two hand weapon as the multiplier is still 1.0.
/🆑